% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/csv_reader.R
\name{csv_reader}
\alias{csv_reader}
\title{reads all CSVs from a folder}
\usage{
csv_reader(directory_name)
}
\arguments{
\item{directory_name}{name of the directory to read}
}
\value{
lsit of dataframe
}
\description{
reads all CSVs from a folder
}
\examples{
\dontrun{
csv_reader("../../path/to/dir")
}
}
