% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw_multiple_name.R
\name{draw_multiple_name}
\alias{draw_multiple_name}
\title{draws the curves of the number of babys born with the names in name_vector}
\usage{
draw_multiple_name(name_vector)
}
\arguments{
\item{name_vector}{a vector of names to plot}
}
\value{
ggplot graph
}
\description{
draws the curves of the number of babys born with the names in name_vector
}
\examples{
\dontrun{
draw_multiple_name( c("Joe", "Kalysie"))
}

}
