% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw_name.R
\name{draw_name}
\alias{draw_name}
\title{draws the number born in France with a specific name and and sex}
\usage{
draw_name(the_name, the_sex)
}
\arguments{
\item{the_name}{name of the babies}

\item{the_sex}{sex of the babies}
}
\value{
ggplot graph
}
\description{
draws the number born in France with a specific name and and sex
}
\examples{
\dontrun{
draw_name("Antoine", "M")
}

}
