% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_as_csv.R
\name{save_as_csv}
\alias{save_as_csv}
\title{Title : function that checks the authorizations of the file before writing a DataFrame into a csv}
\usage{
save_as_csv(data, file_path, ...)
}
\arguments{
\item{data}{dataframe to be writen in a csv file}

\item{file_path}{path to folder}

\item{...}{rest of the parameters of the write.csv2 functions}
}
\value{
file_path
}
\description{
Title : function that checks the authorizations of the file before writing a DataFrame into a csv
}
\examples{
\dontrun{
save_as_csv(MyDataFrame, "../path/to/folder")

MyDataFrame \%>\% save_as_csv("../path/to/folder")
}

}
