% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xlsx_reader.R
\name{xlsx_reader}
\alias{xlsx_reader}
\title{extracts DataFrame from xlsx sheets (one dataframe per sheet)}
\usage{
xlsx_reader(xlsx_path)
}
\arguments{
\item{xlsx_path}{path to xlsx file}
}
\value{
list of dataframe
}
\description{
extracts DataFrame from xlsx sheets (one dataframe per sheet)
}
\examples{
\dontrun{
xlsx_reader(../../path/to/file.xlsx)
}

}
